pyproject.toml
[tool.poetry]
name = "scheduler"
version = "0.1.0"
description = ""
authors = ["ihsgnef <shifeng@cs.umd.edu>"]

[tool.poetry.dependencies]
python = "^3.9"
transformers = "^4.21.0"
SQLAlchemy = "^1.4.39"
altair = "^4.2.0"
uvicorn = "^0.18.2"
alembic = "^1.8.1"
jupyter = "^1.0.0"
pytz = "^2022.1"
tqdm = "^4.64.0"
requests = "^2.28.1"
msgpack = "^1.0.4"
msgpack-numpy = "^0.4.8"
joblib = "^1.1.0"
tables = "^3.7.0"
spacy = "^3.4.1"
fastapi = "^0.79.0"
torch = "^2.0"
psycopg2-binary = "^2.9.3"
psycopg2 = "^2.9.3"
pandas = "^1.4.3"
python-dotenv = "^0.20.0"
sklearn = "^0.0"
wandb = "^0.13.3"
jupyter-http-over-ws = "^0.0.8"
netcal = "^1.3.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
